Pitch, Tempo and Volume using FluidSynth
Summary:
Use of the FluidSynth Library to set the pitch.
The documentations does mention that: 0 << **pitch** << 1024 * 16 - 1
The **pitch** **slider** has been set from 0 to 16
Changing the pitch has immediate effect.

The **tempo** has only been reactivated, it does work but needs the current question to be stopped() and played() again.

Reviewers: sandroandrade

Tags: #minuet

Differential Revision: https://phabricator.kde.org/D2713